+++ src/i18n/locales/en/daily-lesson.json
@@ -52,6 +52,8 @@
   },
   "questionTypes": {
     "experience": "Progress",
-    "xp": "XP"
+    "xp": "XP",
+    "numberOfLessons_singular": "{{number}} lesson",
+    "numberOfLessons_plural": "{{number}} lessons"
   }
 }
