+++ src/routeTree.gen.ts
@@ -19,13 +19,13 @@ import { Route as AuthenticatedDashboardProfileProfileHistoryImport } from "./ro
 import { Route as AuthenticatedDashboardProfileProfileIndexImport } from "./routes/_authenticated/_dashboard/_profile/profile/index";
 import { Route as AuthenticatedDashboardProfileProfileTransactionsImport } from "./routes/_authenticated/_dashboard/_profile/profile/transactions";
 import { Route as AuthenticatedDashboardContentEditorImport } from "./routes/_authenticated/_dashboard/content-editor";
-import { Route as AuthenticatedDashboardDailyLessonDailyLessonIdImport } from "./routes/_authenticated/_dashboard/daily-lesson/$dailyLessonId";
 import { Route as AuthenticatedDashboardDailyLessonIndexImport } from "./routes/_authenticated/_dashboard/daily-lesson/index";
 import { Route as AuthenticatedDashboardPracticeCollectionIdIndexImport } from "./routes/_authenticated/_dashboard/practice/$collectionId/index";
 import { Route as AuthenticatedDashboardPracticeCollectionIdSimulatedTestSimulatedTestIdImport } from "./routes/_authenticated/_dashboard/practice/$collectionId/simulated-test/$simulatedTestId";
 import { Route as AuthenticatedDashboardPracticeIndexImport } from "./routes/_authenticated/_dashboard/practice/index";
 import { Route as AuthenticatedDashboardPracticeSimulatedTestResultImport } from "./routes/_authenticated/_dashboard/practice/simulated-test/result";
 import { Route as AuthenticatedDashboardShopIndexImport } from "./routes/_authenticated/_dashboard/shop/index";
+import { Route as AuthenticatedDailyLessonDailyLessonIdImport } from "./routes/_authenticated/daily-lesson/$dailyLessonId";
 import { Route as AuthenticatedPracticeSimulatedTestIndexImport } from "./routes/_authenticated/practice/simulated-test/index";
 import { Route as AuthenticationImport } from "./routes/_authentication";
 import { Route as AuthenticationForgotPasswordImport } from "./routes/_authentication/forgot-password";
@@ -88,6 +88,13 @@ const AuthenticatedDashboardRoute = AuthenticatedDashboardImport.update({
   getParentRoute: () => AuthenticatedRoute,
 } as any);
 
+const AuthenticatedDailyLessonDailyLessonIdRoute =
+  AuthenticatedDailyLessonDailyLessonIdImport.update({
+    id: "/daily-lesson/$dailyLessonId",
+    path: "/daily-lesson/$dailyLessonId",
+    getParentRoute: () => AuthenticatedRoute,
+  } as any);
+
 const AuthenticatedDashboardContentEditorRoute = AuthenticatedDashboardContentEditorImport.update({
   id: "/content-editor",
   path: "/content-editor",
@@ -122,13 +129,6 @@ const AuthenticatedDashboardDailyLessonIndexRoute =
   AuthenticatedDashboardDailyLessonIndexImport.update({
     id: "/daily-lesson/",
     path: "/daily-lesson/",
-    getParentRoute: () => AuthenticatedDashboardRoute,
-  } as any);
-
-const AuthenticatedDashboardDailyLessonDailyLessonIdRoute =
-  AuthenticatedDashboardDailyLessonDailyLessonIdImport.update({
-    id: "/daily-lesson/$dailyLessonId",
-    path: "/daily-lesson/$dailyLessonId",
     getParentRoute: () => AuthenticatedDashboardRoute,
   } as any);
 
@@ -262,12 +262,12 @@ declare module "@tanstack/react-router" {
       preLoaderRoute: typeof AuthenticatedDashboardContentEditorImport;
       parentRoute: typeof AuthenticatedDashboardImport;
     };
-    "/_authenticated/_dashboard/daily-lesson/$dailyLessonId": {
-      id: "/_authenticated/_dashboard/daily-lesson/$dailyLessonId";
+    "/_authenticated/daily-lesson/$dailyLessonId": {
+      id: "/_authenticated/daily-lesson/$dailyLessonId";
       path: "/daily-lesson/$dailyLessonId";
       fullPath: "/daily-lesson/$dailyLessonId";
-      preLoaderRoute: typeof AuthenticatedDashboardDailyLessonDailyLessonIdImport;
-      parentRoute: typeof AuthenticatedDashboardImport;
+      preLoaderRoute: typeof AuthenticatedDailyLessonDailyLessonIdImport;
+      parentRoute: typeof AuthenticatedImport;
     };
     "/_authenticated/_dashboard/daily-lesson/": {
       id: "/_authenticated/_dashboard/daily-lesson/";
@@ -374,7 +374,6 @@ const AuthenticatedDashboardProfileRouteWithChildren =
 interface AuthenticatedDashboardRouteChildren {
   AuthenticatedDashboardProfileRoute: typeof AuthenticatedDashboardProfileRouteWithChildren;
   AuthenticatedDashboardContentEditorRoute: typeof AuthenticatedDashboardContentEditorRoute;
-  AuthenticatedDashboardDailyLessonDailyLessonIdRoute: typeof AuthenticatedDashboardDailyLessonDailyLessonIdRoute;
   AuthenticatedDashboardDailyLessonIndexRoute: typeof AuthenticatedDashboardDailyLessonIndexRoute;
   AuthenticatedDashboardPracticeIndexRoute: typeof AuthenticatedDashboardPracticeIndexRoute;
   AuthenticatedDashboardShopIndexRoute: typeof AuthenticatedDashboardShopIndexRoute;
@@ -386,8 +385,6 @@ interface AuthenticatedDashboardRouteChildren {
 const AuthenticatedDashboardRouteChildren: AuthenticatedDashboardRouteChildren = {
   AuthenticatedDashboardProfileRoute: AuthenticatedDashboardProfileRouteWithChildren,
   AuthenticatedDashboardContentEditorRoute: AuthenticatedDashboardContentEditorRoute,
-  AuthenticatedDashboardDailyLessonDailyLessonIdRoute:
-    AuthenticatedDashboardDailyLessonDailyLessonIdRoute,
   AuthenticatedDashboardDailyLessonIndexRoute: AuthenticatedDashboardDailyLessonIndexRoute,
   AuthenticatedDashboardPracticeIndexRoute: AuthenticatedDashboardPracticeIndexRoute,
   AuthenticatedDashboardShopIndexRoute: AuthenticatedDashboardShopIndexRoute,
@@ -405,11 +402,13 @@ const AuthenticatedDashboardRouteWithChildren = AuthenticatedDashboardRoute._add
 
 interface AuthenticatedRouteChildren {
   AuthenticatedDashboardRoute: typeof AuthenticatedDashboardRouteWithChildren;
+  AuthenticatedDailyLessonDailyLessonIdRoute: typeof AuthenticatedDailyLessonDailyLessonIdRoute;
   AuthenticatedPracticeSimulatedTestIndexRoute: typeof AuthenticatedPracticeSimulatedTestIndexRoute;
 }
 
 const AuthenticatedRouteChildren: AuthenticatedRouteChildren = {
   AuthenticatedDashboardRoute: AuthenticatedDashboardRouteWithChildren,
+  AuthenticatedDailyLessonDailyLessonIdRoute: AuthenticatedDailyLessonDailyLessonIdRoute,
   AuthenticatedPracticeSimulatedTestIndexRoute: AuthenticatedPracticeSimulatedTestIndexRoute,
 };
 
@@ -446,7 +445,7 @@ export interface FileRoutesByFullPath {
   "/sign-up": typeof AuthenticationSignUpRoute;
   "/verify-otp": typeof AuthenticationVerifyOtpRoute;
   "/content-editor": typeof AuthenticatedDashboardContentEditorRoute;
-  "/daily-lesson/$dailyLessonId": typeof AuthenticatedDashboardDailyLessonDailyLessonIdRoute;
+  "/daily-lesson/$dailyLessonId": typeof AuthenticatedDailyLessonDailyLessonIdRoute;
   "/daily-lesson": typeof AuthenticatedDashboardDailyLessonIndexRoute;
   "/practice": typeof AuthenticatedDashboardPracticeIndexRoute;
   "/shop": typeof AuthenticatedDashboardShopIndexRoute;
@@ -469,7 +468,7 @@ export interface FileRoutesByTo {
   "/sign-up": typeof AuthenticationSignUpRoute;
   "/verify-otp": typeof AuthenticationVerifyOtpRoute;
   "/content-editor": typeof AuthenticatedDashboardContentEditorRoute;
-  "/daily-lesson/$dailyLessonId": typeof AuthenticatedDashboardDailyLessonDailyLessonIdRoute;
+  "/daily-lesson/$dailyLessonId": typeof AuthenticatedDailyLessonDailyLessonIdRoute;
   "/daily-lesson": typeof AuthenticatedDashboardDailyLessonIndexRoute;
   "/practice": typeof AuthenticatedDashboardPracticeIndexRoute;
   "/shop": typeof AuthenticatedDashboardShopIndexRoute;
@@ -496,7 +495,7 @@ export interface FileRoutesById {
   "/_authentication/verify-otp": typeof AuthenticationVerifyOtpRoute;
   "/_authenticated/_dashboard/_profile": typeof AuthenticatedDashboardProfileRouteWithChildren;
   "/_authenticated/_dashboard/content-editor": typeof AuthenticatedDashboardContentEditorRoute;
-  "/_authenticated/_dashboard/daily-lesson/$dailyLessonId": typeof AuthenticatedDashboardDailyLessonDailyLessonIdRoute;
+  "/_authenticated/daily-lesson/$dailyLessonId": typeof AuthenticatedDailyLessonDailyLessonIdRoute;
   "/_authenticated/_dashboard/daily-lesson/": typeof AuthenticatedDashboardDailyLessonIndexRoute;
   "/_authenticated/_dashboard/practice/": typeof AuthenticatedDashboardPracticeIndexRoute;
   "/_authenticated/_dashboard/shop/": typeof AuthenticatedDashboardShopIndexRoute;
@@ -568,7 +567,7 @@ export interface FileRouteTypes {
     | "/_authentication/verify-otp"
     | "/_authenticated/_dashboard/_profile"
     | "/_authenticated/_dashboard/content-editor"
-    | "/_authenticated/_dashboard/daily-lesson/$dailyLessonId"
+    | "/_authenticated/daily-lesson/$dailyLessonId"
     | "/_authenticated/_dashboard/daily-lesson/"
     | "/_authenticated/_dashboard/practice/"
     | "/_authenticated/_dashboard/shop/"
@@ -617,6 +616,7 @@ export const routeTree = rootRoute
       "filePath": "_authenticated.tsx",
       "children": [
         "/_authenticated/_dashboard",
+        "/_authenticated/daily-lesson/$dailyLessonId",
         "/_authenticated/practice/simulated-test/"
       ]
     },
@@ -636,7 +636,6 @@ export const routeTree = rootRoute
       "children": [
         "/_authenticated/_dashboard/_profile",
         "/_authenticated/_dashboard/content-editor",
-        "/_authenticated/_dashboard/daily-lesson/$dailyLessonId",
         "/_authenticated/_dashboard/daily-lesson/",
         "/_authenticated/_dashboard/practice/",
         "/_authenticated/_dashboard/shop/",
@@ -679,9 +678,9 @@ export const routeTree = rootRoute
       "filePath": "_authenticated/_dashboard/content-editor.tsx",
       "parent": "/_authenticated/_dashboard"
     },
-    "/_authenticated/_dashboard/daily-lesson/$dailyLessonId": {
-      "filePath": "_authenticated/_dashboard/daily-lesson/$dailyLessonId.tsx",
-      "parent": "/_authenticated/_dashboard"
+    "/_authenticated/daily-lesson/$dailyLessonId": {
+      "filePath": "_authenticated/daily-lesson/$dailyLessonId.tsx",
+      "parent": "/_authenticated"
     },
     "/_authenticated/_dashboard/daily-lesson/": {
       "filePath": "_authenticated/_dashboard/daily-lesson/index.tsx",
